// web/src/api/model/message.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * school-manager-api
 * OpenAPI spec version: 1.0
 */
import type { ActorTypes } from './actorTypes';
import type { VoiceflowResponse } from './voiceflowResponse';

export interface Message {
    actor: ActorTypes;
    created_at: string;
    is_deleted: boolean;
    parent_text?: string;
    voiceflowResponses?: VoiceflowResponse[];
}