// web/src/pam_api/model/actorTypes.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * pam-ai-whatsapp
 * OpenAPI spec version: 1.0.0
 */

export type ActorTypes = (typeof ActorTypes)[keyof typeof ActorTypes];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ActorTypes = {
    User: 'User',
    Assistant: 'Assistant',
} as const;